requirements.txt
transformers
torch 
torchvision 
torchaudio
pypdf
python-docx
langdetect 
sentencepiece 
sacremoses